Combine all python-forge modules and license into one file
Also apply formatting, remove unused type-ignores, and fix misc typing issues
