--- pyinaturalist/_vendor/forge/__init__.py
@@ -1,44 +1,0 @@
-from ._config import (
-    get_run_validators,
-    set_run_validators,
-)
-from ._exceptions import (
-    ForgeError,
-    ImmutableInstanceError,
-)
-from ._marker import (
-    empty,
-    void,
-)
-from ._revision import (
-    Mapper,
-    Revision,
-    # Group
-    compose,
-    copy,
-    manage,
-    returns,
-    sort,
-    synthesize, sign,
-    # Unit
-    delete,
-    insert,
-    modify,
-    replace,
-    translocate, move,
-)
-from ._signature import (
-    Factory,
-    FParameter,
-    FSignature,
-    findparam,
-    fsignature,
-    pos, pok, vpo, kwo, vkw,
-    arg, ctx, args, kwarg, kwargs,
-    self_ as self,
-    cls_ as cls,
-)
-from ._utils import (
-    callwith,
-    repr_callable,
-)
