+++ pyproject.toml
@@ -132,7 +132,6 @@ skip_gitignore = true
 known_first_party = ['test']
 
 [tool.mypy]
-check-untyped-defs = true
 ignore_missing_imports = true
 pretty = true
 show_column_numbers = true
